Refresh haiku on cafe page when generated
Prior to this change, the face page needed to be reloaded to see the
latest haiku

This change updates the HTML after the poem has been generated. Post
saving javascript a call to refresh the haiku gets the newest version and displays
it. This also required a new view and a new url for the view.